web/requirements.txt
Django==3.1.1
djangorestframework==3.11.1
gunicorn==19.9.0
psycopg2==2.8.5
Pillow
django-filter==2.4.0